+++ StatelessCMS.php
@@ -3,7 +3,7 @@
  * StatelessCMS.php
  * 
  * @package StatelessCMS
- * @version 1.0.0
+ * @version 1.1.0
  */
 
 /**